refactor: AppBar를 page 폴더에 분류
