+++ src/components/page/AppBar.tsx
@@ -1,12 +1,12 @@
 import { useRecoilState } from 'recoil';
 import { styled } from 'styled-components';
+import servers from '../../config/servers';
+import serverState from '../../recoil/atoms/serverState';
+import userCartItemsState from '../../recoil/user/userCartItemsState';
+import userProfileState from '../../recoil/user/userProfileState';
 import HomeIcon from '../assets/icons/home-icon.svg';
-import servers from '../config/servers';
-import serverState from '../recoil/atoms/serverState';
-import userCartItemsState from '../recoil/user/userCartItemsState';
-import userProfileState from '../recoil/user/userProfileState';
-import Badge from './common/Badge';
-import AwaitRecoilState from './utils/AwaitRecoilState';
+import Badge from '../common/Badge';
+import AwaitRecoilState from '../utils/AwaitRecoilState';
 
 const AppBarContainer = styled.header`
   width: 100%;
